lavillasite/server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.0.6
Flask-Migrate==4.1.0
